+++ libdino/src/service/database.vala
@@ -16,7 +16,7 @@ using Dino.Entities;
 namespace Dino {
 
 public class Database : Qlite.Database {
-    private const int VERSION = 33;
+    private const int VERSION = 34;
 
     public class AccountTable : Table {
         public Column<int> id = new Column.Integer("id") { primary_key = true, auto_increment = true };
@@ -320,10 +320,11 @@ public class Database : Qlite.Database {
         public Column<int> send_marker = new Column.Integer("send_marker") { min_version=3 };
         public Column<int> pinned = new Column.Integer("pinned") { default="0", min_version=25 };
         public Column<long> history_cleared_at = new Column.Long("history_cleared_at") { default="0", min_version=32 };
+        public Column<int> message_expiry_seconds = new Column.Integer("message_expiry_seconds") { default="0", min_version=34 };
 
         internal ConversationTable(Database db) {
             base(db, "conversation");
-            init({id, account_id, jid_id, resource, active, active_last_changed, last_active, type_, encryption, read_up_to, read_up_to_item, notification, send_typing, send_marker, pinned, history_cleared_at});
+            init({id, account_id, jid_id, resource, active, active_last_changed, last_active, type_, encryption, read_up_to, read_up_to_item, notification, send_typing, send_marker, pinned, history_cleared_at, message_expiry_seconds});
         }
     }