+++ main/src/ui/conversation_content_view/conversation_view.vala
@@ -40,6 +40,7 @@ public class ConversationView : Widget, Plugins.ConversationItemCollection, Plug
     private Gee.List<Widget> widget_order = new Gee.ArrayList<Widget>();
     private ContentProvider content_populator;
     private SubscriptionNotitication subscription_notification;
+    private ExpiryNotification expiry_notification;
 
     private double? was_value;
     private double? was_upper;
@@ -124,6 +125,7 @@ public class ConversationView : Widget, Plugins.ConversationItemCollection, Plug
 
         content_populator = new ContentProvider(stream_interactor);
         subscription_notification = new SubscriptionNotitication(stream_interactor);
+        expiry_notification = new ExpiryNotification(stream_interactor);
 
         add_meta_notification.connect(on_add_meta_notification);
         remove_meta_notification.connect(on_remove_meta_notification);
@@ -434,6 +436,7 @@ public class ConversationView : Widget, Plugins.ConversationItemCollection, Plug
         }
         content_populator.init(this, conversation, Plugins.WidgetType.GTK4);
         subscription_notification.init(conversation, this);
+        expiry_notification.init(conversation, this);
     }
 
     private void display_latest() {